+++ OptimoveSDK/Tests/Sources/Events/CoreEvents/NotificationOpenedEventTests.swift
@@ -24,7 +24,7 @@ class NotificationOpenedEventTests: XCTestCase {
         XCTAssertEqual(event.name, campaign.eventName)
         XCTAssertEqual(event.context[NotificationOpenedEvent.Constants.Key.appNS] as? String , bundleIdentifier)
         XCTAssertEqual(event.context[NotificationOpenedEvent.Constants.Key.identityToken] as? String, identityToken)
-        XCTAssertEqual(event.context[NotificationOpenedEvent.Constants.Key.timestamp] as? Int, timestamp.timeIntervalSince1970.seconds)
+        XCTAssertEqual(event.context[NotificationOpenedEvent.Constants.Key.timestamp] as? Int64, timestamp.timeIntervalSince1970.seconds)
     }
 
     func test_triggered_opened_event() {
@@ -45,7 +45,7 @@ class NotificationOpenedEventTests: XCTestCase {
         XCTAssertEqual(event.name, campaign.eventName)
         XCTAssertEqual(event.context[NotificationOpenedEvent.Constants.Key.appNS] as? String , bundleIdentifier)
         XCTAssertEqual(event.context[NotificationOpenedEvent.Constants.Key.identityToken] as? String, identityToken)
-        XCTAssertEqual(event.context[NotificationOpenedEvent.Constants.Key.timestamp] as? Int, timestamp.timeIntervalSince1970.seconds)
+        XCTAssertEqual(event.context[NotificationOpenedEvent.Constants.Key.timestamp] as? Int64, timestamp.timeIntervalSince1970.seconds)
     }
 
 }
